backend/pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "RAG backend service"
requires-python = ">=3.13"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "numpy",
  "sqlalchemy",
  "alembic",
  "psycopg[binary]",
  "pgvector",
  "bcrypt",
  "python-jose[cryptography]",
  "sentence-transformers",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "httpx",
]
llm = [
  "llama-cpp-python",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
